tests/wpt/tests/speculation-rules/speculation-measurement/performance-speculations-preredirect-urls.tentative.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>performance.getSpeculations() - URLs are pre-redirect request URLs</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/utils.js"></script>
<script src="/speculation-rules/resources/utils.js"></script>
<script src="/speculation-rules/prefetch/resources/utils.sub.js"></script>
<script src="support/speculation-measurement-utils.js"></script>
<body>
<script>
setup(() => {
  assert_true(isSpeculationMeasurementEnabled(),
    "SpeculationMeasurement feature must be enabled");
});

// Returns { redirectUrl, destinationUrl } where redirectUrl 302s to
// destinationUrl. `tag` cache-busts between tests.
function makeRedirectChain(tag, destinationPath) {
  const destinationUrl =
      new URL(`${destinationPath}?${tag}-dest`, location.href).href;
  const redirectUrl = new URL(
      `/common/redirect.py?tag=${tag}` +
      `&location=${encodeURIComponent(destinationUrl)}`,
      location.href).href;
  return { redirectUrl, destinationUrl };
}

// Preloads: exposed URL must be the request URL, not the post-redirect URL.

promise_test(async t => {
  const { redirectUrl, destinationUrl } = makeRedirectChain(
      'preload-redirect-script', 'support/preloaded-script.js');

  await addPreloadAndWait({t, as: 'script', href: redirectUrl});

  const preloads = performance.getSpeculations().preloads;

  const preRedirectEntry = preloads.find(p => p.url === redirectUrl);
  assert_not_equals(preRedirectEntry, undefined,
    `An entry keyed by the request URL (${redirectUrl}) must exist`);
  assert_equals(preRedirectEntry.as, 'script', "as should be 'script'");

  const postRedirectEntry = preloads.find(p => p.url === destinationUrl);
  assert_equals(postRedirectEntry, undefined,
    `No entry should be keyed by the post-redirect URL (${destinationUrl})`);
}, "preloads[].url is the request URL, not the post-redirect URL");

promise_test(async t => {
  // Same property after the preload is consumed.
  const { redirectUrl, destinationUrl } = makeRedirectChain(
      'preload-redirect-used', 'support/preloaded-script.js');

  await addPreloadAndWait({t, as: 'script', href: redirectUrl});
  await usePreload({t, as: 'script', href: redirectUrl});

  const preloads = performance.getSpeculations().preloads;
  const entry = preloads.find(p => p.url === redirectUrl);
  assert_not_equals(entry, undefined,
    "Used preload should still be keyed by the request URL");
  assert_not_equals(entry.used, null,
    "Preload should be marked used");

  assert_equals(preloads.find(p => p.url === destinationUrl), undefined,
    "No entry should be keyed by the post-redirect URL after use either");
}, "preloads[].url stays pre-redirect after the preload is consumed");

// Speculation rules: exposed URL must be the rule URL, not any URL the
// browser may have followed during prefetch/prerender execution.

promise_test(async t => {
  const { redirectUrl, destinationUrl } = makeRedirectChain(
      'sr-prefetch-redirect', '/common/blank.html');

  insertSpeculationRules({
    prefetch: [{
      urls: [redirectUrl],
      eagerness: 'eager',
      tag: 'preredirect-prefetch-test',
    }],
  });

  await new Promise(resolve => t.step_timeout(resolve, 100));

  const navigations = performance.getSpeculations().navigations;

  const preRedirectEntry = navigations.find(n => n.url === redirectUrl);
  assert_not_equals(preRedirectEntry, undefined,
    `An entry keyed by the rule URL (${redirectUrl}) must exist`);
  assert_equals(preRedirectEntry.type, 'prefetch');
  assert_equals(preRedirectEntry.eagerness, 'eager');

  const postRedirectEntry = navigations.find(n => n.url === destinationUrl);
  assert_equals(postRedirectEntry, undefined,
    `No entry should be keyed by the post-redirect URL (${destinationUrl})`);
}, "navigations[].url for prefetch is the rule URL, not the post-redirect URL");

promise_test(async t => {
  const { redirectUrl, destinationUrl } = makeRedirectChain(
      'sr-prerender-redirect', '/common/blank.html');

  insertSpeculationRules({
    prerender: [{
      urls: [redirectUrl],
      eagerness: 'eager',
      tag: 'preredirect-prerender-test',
    }],
  });

  await new Promise(resolve => t.step_timeout(resolve, 100));

  const navigations = performance.getSpeculations().navigations;

  const preRedirectEntry = navigations.find(n => n.url === redirectUrl);
  assert_not_equals(preRedirectEntry, undefined,
    `An entry keyed by the rule URL (${redirectUrl}) must exist`);
  assert_equals(preRedirectEntry.type, 'prerender');

  const postRedirectEntry = navigations.find(n => n.url === destinationUrl);
  assert_equals(postRedirectEntry, undefined,
    `No entry should be keyed by the post-redirect URL (${destinationUrl})`);
}, "navigations[].url for prerender is the rule URL, not the post-redirect URL");
</script>
</body>
